refactor: renamed 'RBTSearchTests.kt' to 'RedBlackSearchTreeTests.kt'
